requirements.txt
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
google-auth==2.15.0
google-auth-oauthlib==0.8.0
gspread==5.7.2
httplib2==0.21.0
idna==3.4
oauth2client==4.1.3
oauthlib==3.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dotenv==0.21.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
urllib3==1.26.13
